Conflicts:
	rv-predict-agent/src/main/java/rvpredict/logging/LoggingEngine.java
	rv-predict-engine/src/main/java/rvpredict/db/EventOutputStream.java

+++ rv-predict-agent/src/main/java/rvpredict/logging/LoggingEngine.java
@@ -132,7 +132,7 @@ public class LoggingEngine {
     /**
      * Logs an {@link rvpredict.db.EventItem} to the trace.
      *
-     * @see rvpredict.db.EventItem#EventItem(long, long, int, long, long, long, rvpredict.trace.EventType)
+     * @see rvpredict.db.EventItem#EventItem(long, long, int, long, int, long, rvpredict.trace.EventType)
      *      for a more elaborate description of the parameters.
      *
      * @param eventType  type of event being recorded
@@ -141,7 +141,7 @@ public class LoggingEngine {
      * @param addrr additional information identifying the event
      * @param value data involved in the event
      */
-    public void saveEvent(EventType eventType, int id, long addrl, long addrr, long value) {
+    public void saveEvent(EventType eventType, int id, long addrl, int addrr, long value) {
         if (shutdown || skipSavingEvent()) return;
 
         long gid = globalEventID.incrementAndGet();
@@ -151,7 +151,7 @@ public class LoggingEngine {
     }
 
     /**
-     * Wrapper for {@link #saveEvent(rvpredict.trace.EventType, int, long, long, long)}
+     * Wrapper for {@link #saveEvent(rvpredict.trace.EventType, int, long, int, long)}
      * The missing arguments default to 0.
      */
     public void saveEvent(EventType eventType, int locId, long arg) {
@@ -159,7 +159,7 @@ public class LoggingEngine {
     }
 
     /**
-     * Wrapper for {@link #saveEvent(rvpredict.trace.EventType, int, long, long, long)}
+     * Wrapper for {@link #saveEvent(rvpredict.trace.EventType, int, long, int, long)}
      * The missing arguments default to 0.
      */
     public void saveEvent(EventType eventType, int locId) {
