+++ rv-predict-engine/src/main/java/rvpredict/db/EventInputStream.java
@@ -34,7 +34,7 @@ public class EventInputStream extends DataInputStream {
                 readLong(),
                 readInt(),
                 readLong(),
-                readLong(),
+                readInt(),
                 readLong(),
                 EventType.values()[readByte()]);
     }
